feat: dedicated project for examples, bump version
