+++ example/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "smart-handles-sdk-example",
+  "version": "0.2.0",
+  "description": "A set of predefined scenarios for making swaps on preprod testnet",
+  "main": "./dist/index.js",
+  "type": "module",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsup src/index.ts --minify --format esm",
+    "quick-test": "tsup src/index.ts --format esm && node dist/index.js --help"
+  },
+  "author": "Anastasia Labs",
+  "license": "ISC",
+  "devDependencies": {
+    "@types/node": "^20.12.12",
+    "tsup": "^8.0.2",
+    "typescript": "^5.4.5"
+  },
+  "dependencies": {
+    "@anastasia-labs/smart-handles-offchain": "^0.2.0",
+    "chalk": "^5.3.0",
+    "commander": "^12.1.0"
+  }
+}
