+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@anastasia-labs/smart-handles-offchain",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "https://docs.github.com/en/packages/quickstart",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -38,5 +38,6 @@
     "@anastasia-labs/lucid-cardano-fork": "^0.10.8-rc.3",
     "@blockfrost/blockfrost-js": "^5.5.0",
     "@minswap/sdk": "^0.2.2-beta"
-  }
+  },
+  "packageManager": "pnpm@9.0.4+sha1.b198ac6d38244fd829253720f9daafd6a606834d"
 }
